feat(revampedtradenavigation): revampedTradeNavigation
RevampedTradeNavigation
