+++ client/trademon/src/components/containerComponents/offerDetailsPokemon.tsx
@@ -1,15 +1,14 @@
 import React, { useEffect, useState, Fragment } from 'react';
 import '../../styling/containers.scss';
 import NavComponent from '../navComponents/navComponent';
-import ContactSeller from './contactSeller';
+import ContactSeller from '../aaPageComponents/contactSeller';
 import { Trades } from '../../utils/interfaces';
 import { getOnePokeTrade, getUserPublicDetails } from '../../utils/rest';
 import { useParams } from 'react-router';
 import setAppraisalImage, { calcRating } from '../../utils/helperFunctions';
 import { PuffLoader } from 'react-spinners';
-import TabsComponent from '../navComponents/tabsComponent';
-    
-export default function OfferDetailsPage() {
+ 
+export default function OfferDetailsPoke() {
   const { tradeID }: any = useParams();
   const [appraisalImgUrl, setAppraisalImgUrl] = useState<string>('');
   const [sellerRatingValue, setNumSellerRatingValue] = useState<number>(0);
@@ -19,7 +18,6 @@ export default function OfferDetailsPage() {
   const [userPublicDetails, setUserPublicDetails] = useState<any>({
     numberOfSales: 0,
   });
-  const [detailsTchatF, setdetailsTchatF] = useState<boolean>(true);
 
   const [tradeDetails, setTradeDetails] = useState<Trades>({
     id: 0,
@@ -125,9 +123,7 @@ export default function OfferDetailsPage() {
   };
 
   return (
-    <div className="offer-details-page">
-      <NavComponent></NavComponent>
-      <TabsComponent detailsTchatF={detailsTchatF}></TabsComponent>
+    <div>
       <div className="offer-details-container">
         <div className="item-details-container">
           <div className="small-text">ID: {tradeDetails!.id}</div>
