+++ client/trademon/src/components/tileComponents/userTile.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../styling/temp_tradePage.scss';
+import '../../styling/temp_chatPage.scss';
 import imageSeller from '../../assets/mtgs/1.png';
 
 export default function UserTile() {
