+++ client/trademon/src/styling/containers.scss
@@ -311,7 +311,7 @@
 
 .interest-box {
   @extend .flex-column-center;
-  width: 60%;
+  width: 90vw;
   flex-wrap: wrap;
   margin-top: 10px;
 }
