--- client/trademon/src/styling/temp_tradePage.scss
@@ -1,228 +1,0 @@
-@import './config';
-@import './animations';
-
-.trading-container {
-  height: 100vh;
-  width: 100%;
-  background-image: linear-gradient(
-      to right bottom,
-      rgba(0, 0, 0, 0.9),
-      rgba(0, 0, 0, 0.8)
-    ),
-    url(../assets/bg_trade_MTGO.jpg);
-  background-size: cover;
-  background-position: center;
-  text-align: center;
-  overflow: auto;
-}
-
-.parent {
-  @extend .flex-column-center;
-  width: 100%;
-  height: 100vh;
-}
-
-/* 1<--->2<--->3<--->4 */
-.trading-container-header {
-  font-size: 2.6rem;
-  display: flex;
-  justify-content: space-between;
-  width: 100vw;
-  align-items: center;
-  text-align: justify;
-}
-
-.title-spacer {
-  width: 10vw;
-}
-.trade-room-title-text {
-  font-size: calc(14px + 3.5vw);
-}
-
-.master-trade-container {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-wrap: wrap;
-  height: 90vh;
-}
-
-.left-side {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  max-height: 90vh;
-  width: 49vw;
-  min-width: 370px;
-  margin-top: 2vh;
-}
-
-.traders-container {
-  @extend .flex-center;
-  margin: 2vw 2vw;
-  width: 100%;
-}
-
-.trader-container {
-  @extend .flex-column-center;
-  margin: 0vw 2vw;
-  width: 100%;
-}
-
-.user_image {
-  width: 10vw;
-  height: 10vw;
-  min-width: 50px;
-  min-height: 50px;
-  image-rendering: crisp-edges;
-}
-
-.user-text-topp {
-  font-size: calc(10px + 1.2vw);
-  width: 20vw;
-  min-width: 175px;
-  margin: 1vh 0vh;
-  background-color: $colorGreyPop;
-  color: white;
-  padding: 1vh 2vh;
-  border-radius: 25px;
-}
-
-.user-badges {
-  @extend .flex-center;
-  width: 100%;
-}
-
-.user-badge {
-  width: 3vw;
-  height: 3vw;
-  min-width: 30px;
-  min-height: 30px;
-  image-rendering: pixelated;
-}
-
-.right-side {
-  height: 90%;
-  width: 49vw;
-  min-width: 370px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: flex-start;
-  margin-top: 2vh;
-}
-
-.chat-container {
-  width: 95%;
-  max-height: 50vh;
-  min-height: 50vh;
-  overflow-y: auto;
-  background-color: $colorGreyPop80;
-  border-radius: 25px;
-  padding: 1vw;
-}
-
-.trade-item-size {
-  width: 12vw;
-  height: 18vw;
-  min-width: 180px;
-  min-height: 250px;
-  animation: floatnoglow 4s ease-in-out infinite;
-}
-
-.trade-room-details {
-  display: flex;
-  justify-content: space-around;
-  width: 80%;
-  img {
-    margin-top: 1em;
-    animation: floatnoglow 4s ease-in-out infinite;
-  }
-  .trade-room-details-text {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-  }
-}
-
-.trade-item-size-square {
-  width: 20vw;
-  height: 20vw;
-  min-width: 200px;
-  min-height: 300px;
-}
-
-.traderoom-trade-details {
-  width: 100%;
-  display: flex;
-  justify-content: space-around;
-  border: 1px solid white;
-}
-
-.user-comm {
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  max-width: 600px;
-  width: 75vw;
-  min-width: 100%;
-}
-
-.user-textarea {
-  background-color: $colorGreyPop80;
-  margin-bottom: 10px;
-  border-radius: 25px;
-}
-
-.trade-button-row {
-  display: flex;
-}
-
-.dm-offer-buttons {
-  width: 100%;
-  margin: 0vw 1vw;
-}
-.offer-trade-button {
-  background-color: $colorSecondary;
-}
-
-.offer-trade-button:hover {
-  color: $colorSecondary;
-}
-
-.trade-room-nav {
-  width: 100vw;
-}
-
-.nav-container-smallerr {
-  display: flex;
-  align-items: center;
-  position: relative;
-  max-height: 9vh;
-  padding-bottom: 1em;
-  margin-bottom: 1em;
-  border-bottom: 2px solid $colorPrimary;
-  box-shadow: 0 0 10px 3px $colorPrimary;
-}
-
-.logo-img-link-smaller {
-  width: 5vw;
-  height: 5vw;
-  min-width: 55px;
-  min-height: 55px;
-  padding-left: 15px;
-  padding-top: 2vh;
-}
-
-.logo-img-smaller {
-  width: 100%;
-  opacity: 90%;
-  cursor: pointer;
-}
-
-.trademon-title-link-smaller {
-  cursor: pointer;
-  max-width: 15vw;
-  justify-self: flex-start;
-}
